将Effect坐标统一为int

+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect1641.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect1641 extends Effect {
 
-    public Effect1641(double x, double y, double scale) {
+    public Effect1641(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
