+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect1644.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect1644 extends Effect {
 
-    public Effect1644(double x, double y, double scale) {
+    public Effect1644(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
