+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect1647.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect1647 extends Effect {
 
-    public Effect1647(double x, double y, double scale) {
+    public Effect1647(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
