+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect1649.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect1649 extends Effect {
 
-    public Effect1649(double x, double y, double scale) {
+    public Effect1649(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
