+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect1652.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect1652 extends Effect {
 
-    public Effect1652(double x, double y, double scale) {
+    public Effect1652(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
