+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect515.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect515 extends Effect {
 
-    public Effect515(double x, double y, double scale) {
+    public Effect515(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
