+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect519.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect519 extends Effect {
 
-    public Effect519(double x, double y, double scale) {
+    public Effect519(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
