+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect520.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect520 extends Effect {
 
-    public Effect520(double x, double y, double scale) {
+    public Effect520(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
