+++ src/main/java/com/github/jiangwangyang/square/adventure/cell/effect/Effect522.java
@@ -2,7 +2,7 @@ package com.github.jiangwangyang.square.adventure.cell.effect;
 
 public class Effect522 extends Effect {
 
-    public Effect522(double x, double y, double scale) {
+    public Effect522(int x, int y, double scale) {
         this.x = x;
         this.y = y;
         this.scale = scale;
